requirements.txt
streamlit>=1.35.0
akshare>=1.10.95
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=6.5.0
scikit-learn>=1.2.0
scipy>=1.10.0
requests>=2.31.0
python-dateutil>=2.8.2
openai>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
numba>=0.58.0
tushare>=1.4.0